refactor: split code to separate files
