+++ .eslintrc.json
@@ -10,6 +10,7 @@
         "ecmaVersion": "latest",
         "sourceType": "module"
     },
+    "parser": "babel-eslint",
     "rules": {
         "linebreak-style": 0,
         "global-require": 0,
