+++ index.html
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-  <script src="./index.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 
 </html>
